Trying to squish some button bugs...
I have no idea if it's code related or if it's wiring related, but sometimes some of the buttons don't work. The rotary encoder *always* works, as does the new band change button I just added. The routines for those buttons are basically the same as the ones for the other buttons that sometimes don't work. I'll have to tone everything out and make sure the connections are sound. I think the code looks OK considering what DOES work.