Update 328VFO.ino
I'm starting to try and separate out the button IO so it's more modular and not quite as hard-coded. So encoder_left() and encoder_right(); are a first crack at that. I will still have to put lots of conditionals in but at least they're an indent back now, and the code is clearer to read.