Update 328VFO.ino
Changing this to be a bit more logical. The chip works in fractions of a hz and you can multiply by 100ULL to get hz. But there’s a define you can use instead so the code reads better. You also don’t need the ULLs in the band starts. It’s a wonder it ever worked before.